Landscape friendly mobile display and more.
Added slight shadow on the bottom of the header area in order to create
a division between elements on the body and the header itself on scroll.
Adjusted settings for header divider on mobile screens.
Resized header areas on mobile for more screen visibility.
Added landscape mobile-friendly mode for screens of less than 736v/420h.